requirements.txt
pyperclip==1.8.2
streamlit==1.33.0
openai==1.23.2
python-dotenv==1.0.1
